+++ LocalPackages/net.peeweek.gameplay-ingredients/Runtime/Discover/Discover.cs
@@ -11,11 +11,12 @@ public class Discover : MonoBehaviour
     public bool DefaultSelected = false;
     public bool AlignViewToTransform = false;
 
-    public GameObjectToggle[] ObjectsToToggle;
+    public GameObjectToggle[] ObjectsToToggle = new GameObjectToggle[0];
     [Multiline]
     public string Description = "Some Description of the Component\n\nCan be set as multiple lines.";
+    public string Tags = "";
     public int Priority = 0;
-    public DiscoverSection[] Sections;
+    public DiscoverSection[] Sections = new DiscoverSection[0];
 
 #if UNITY_EDITOR
     [UnityEditor.MenuItem("GameObject/Gameplay Ingredients/Discover", priority=10)]